// src/shared/api/generated/operation/authTokensResponse.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Amazing EKB Hub API
 * MVP backend API (OpenAPI-first)
 * OpenAPI spec version: 1.0.0
 */

/**
 * Контракт успешной аутентификации.
 */
export type AuthTokensResponse = {
  /** JWT access token для авторизованных запросов. */
  accessToken: string;
  /** Refresh token для получения новой пары токенов. */
  refreshToken: string;
  /** Тип токена для заголовка `Authorization`. */
  tokenType: string;
  /** TTL access token в человекочитаемом формате. */
  accessExpiresIn?: string;
};